// packages/vitest/src/node/pools/types.ts
import type { Awaitable } from '@vitest/utils'
import type { ContextRPC } from '../../types/worker'
import type { TestProject } from '../project'
import type { TestSpecification } from '../spec'

export interface ProcessPool {
  name: string
  runTests: (files: TestSpecification[], invalidates?: string[]) => Awaitable<void>
  collectTests: (files: TestSpecification[], invalidates?: string[]) => Awaitable<void>
  close?: () => Awaitable<void>
}

export interface Runtime {
  name: string
  isStarted: boolean
  reportMemory?: boolean
  options: {
    distPath: string
    project: TestProject
    workerId: number
    method: 'run' | 'collect'
    cacheFs?: boolean
  }

  /** Note that start can be called multiple times. First time indicates worker warmup. */
  start: (options: Pick<Task, 'env' | 'execArgv'>) => Promise<void>
  stop: () => Promise<void>
  on: ((event: 'message', callback: (message: WorkerResponse) => void) => void) & ((event: 'error', callback: (error: Error) => void) => void)
  off: ((event: 'message', callback: (message: WorkerResponse) => void) => void) & ((event: 'error', callback: (error: Error) => void) => void)
  onWorker: (event: string, callback: (arg: any) => void) => void
  offWorker: (event: string, callback: (arg: any) => void) => void
  postMessage: (message: WorkerRequest) => void
  serialize: (message: any) => any
  deserialize: (message: any) => any
}

export interface Task {
  runtime: 'forks' | 'threads' | 'vmForks' | 'vmThreads' | (string & {})
  project: TestProject
  isolate: boolean
  env: Record<string, string>
  execArgv: string[]
  context: ContextRPC
  memoryLimit: number | null
}

export type WorkerRequest
  = { __vitest_worker_request__: true } & (
    | { type: 'start'; options: { workerFilename: string; workerId: number; reportMemory: boolean } }
    | { type: 'stop' }
    | { type: 'run'; context: ContextRPC }
    | { type: 'collect'; context: ContextRPC }
    | { type: 'cancel' }
)

export type WorkerResponse
  = { __vitest_worker_response__: true } & (
    | { type: 'started' }
    | { type: 'stopped' }
    | { type: 'testfileFinished'; usedMemory?: number }
)